Reorganize Decoded monadic instance
- Implement `map`, `apply`, `flatMap` as instance methods. This mimics
  the behavior found elsewhere.
- Move the operator definitions into Decoded.swift for visibility
